refactor: change project structure for docker
