+++ actions/requirements-actions.txt
@@ -1,0 +1,1 @@
+SQLAlchemy==1.3.19
